Implementation of Repository Structure in the project.
The end goal of this change is to make it possible to refactor the Controller classes and the Service classes. Currently there is too much code being executed in the Controllers.

-Generic IRepository and implementation of Generic IRepository in Repository.
- Models excluding Template classes currently have their own Repository class in order to maintain SoC (seperation).

Created Repositories following Unit of Work Pattern which makes it easier to handle any changes needed to the database. It will collect everything one does which can affect the database. When its finished, this will be provided to be done to change DB as a result of the work.

It is